src/django_teach/requirements.txt
django
python-dotenv
psycopg2-binary
django-countries
djangorestframework
djangorestframework-simplejwt
